remove struct for OoO in difftest.h; many uint64 to uint32

+++ src/test/csrc/common/axi4.h
@@ -24,10 +24,10 @@
 // 4*64 bits
 #define AXI_DATA_WIDTH_64 4
 
-typedef uint64_t axi_addr_t;
-typedef uint64_t axi_data_t[AXI_DATA_WIDTH_64];
+typedef uint32_t axi_addr_t;
+typedef uint32_t axi_data_t[AXI_DATA_WIDTH_64];
 #define axi_copy_data(dest, src) \
-  memcpy(dest, src, sizeof(uint64_t)*AXI_DATA_WIDTH_64);
+  memcpy(dest, src, sizeof(uint32_t)*AXI_DATA_WIDTH_64);
 
 struct axi_aw_channel {
   uint8_t       ready;
